Use CGFloat where necessary.

+++ Demo/Classes/HudDemoViewController.m
@@ -184,7 +184,7 @@
 
 - (void)myProgressTask {
     // This just increases the progress indicator in a loop
-    float progress = 0.0f;
+    CGFloat progress = 0.0f;
     while (progress < 1.0f) {
         progress += 0.01f;
         HUD.progress = progress;
@@ -198,7 +198,7 @@
     // Switch to determinate mode
     HUD.mode = MBProgressHUDModeDeterminate;
     HUD.labelText = @"Progress";
-    float progress = 0.0f;
+    CGFloat progress = 0.0f;
     while (progress < 1.0f)
     {
         progress += 0.01f;
